Public Documentation

Documentation for MimiqCircuitsBase.jl's public interface.

See the Internals section of the manual for internal package docs.

MimiqCircuitsBase

Quantum Circuits and Instructions

MimiqCircuitsBase.Instruction — Type

struct Instruction{N,M,T<:Operation}

Element of a quantum circuit, representing a N-qubit gate applied to N targets

Parameters

gate::T actual gate represented

qtargets::NTuple{N, Int64} indices specifying the quantum bits on which the instruction is applied

ctargets::NTuple{N, Int64} indices specifying the classical bits on which the instruction is applied

MimiqCircuitsBase.getbit — Function

getbit(instruction, i)

Returns the i-th target classical bit of an instruction.

See also getbits, getqubit, getqubits,

MimiqCircuitsBase.getbits — Function

getbits(instruction)

Returns all the classical bits to which the instruction is applied.

See also getbit, getqubits, getqubit,

MimiqCircuitsBase.getoperation — Method

getoperation(instruction)

Returns the quantum operation associated to the given instruction.

MimiqCircuitsBase.getqubit — Function

getqubit(instruction, i)

Returns the i-th target qubit of an instruction.

See also getqubits, getbit, getbits,

MimiqCircuitsBase.getqubits — Function

getqubits(instruction)

Returns all the quantum bits to which the instruction is applied.

See also getqubit, getbits, getbit,

MimiqCircuitsBase.Circuit — Type

struct Circuit

Representation of a quantum circuit as a vector of gates applied to the qubits.

Parameters

gates::Vector{Instruction} vector of quantum instructions (see

Instruction)

Examples

Adding operations to a circuit

Operation can be added one by one to a circuit with the push!(circuit, operation, targets...) function

c = Circuit()
push!(c, GateH(), 1)
push!(c, GateCX(), 1, 2)
push!(c, GateRX(π / 4), 1)
push!(c, Barrier, 1, 3)
push!(c, Measure(), 1, 1)

Targets are not restricted to be single values, but also vectors. In this case a single push! will add multiple operations. The behaviour is similar to Iterators.product. For example:

push!(c, GateCX(), 1, 2:4)

is equivalent to

for targets in Iterators.product(1, 2:4)
    push!(c, GateCX(), targets...)
end

For example

c = Circuit()
push!(c, GateH(),1:3)
push!(c,GateCX(),1:3,4:6)

Some operations behave a bit differently. See also: Barrier and Measure

Available operations

Gates

Single qubit gates

GateX GateY GateZ GateH GateS GateSDG GateT GateTDG GateSX GateSXDG GateID

Single qubit gates (parametric)

GateRX GateRY GateRZ GateP GateR GateU

Two qubit gates

GateCX GateCY GateCZ GateCH GateSWAP GateISWAP, GateISWAPDG

Two qubit gates (parametric)

GateCP GateCRX GateCRY GateCRZ GateCU

Other

GateCustom

No-ops

Barrier

Non-unitary operations

Reset Measure

Composite operations

Control Parallel IfStatement

MimiqCircuitsBase.depth — Method

depth(circuit)

Compute the depth of a quantum circuit.

The depth of a quantum circuit is a metric computing the maximum time (in units of quantum gates application) between the input and output of the circuit.

Operations

MimiqCircuitsBase.Operation — Type

abstract type Operation{N,M}

Parameters

N: number of qubits the operation applies on.

M: number of bits the operation applies on.

Methods

opname

inverse

MimiqCircuitsBase.hilbertspacedim — Function

hilbertspacedim(gate)
hilberspacedim(circuit)

Hilbert space dimension for the given operation.

MimiqCircuitsBase.inverse — Function

inverse(circuit)
inverse(instruction)
inverse(operation)

Return the inverse of the given operation.

MimiqCircuitsBase.numbits — Function

numbits(instruction)
numbits(circuit)

Number of classical bits on which the given operation or instruction is defined.

MimiqCircuitsBase.numqubits — Function

numqubits(gate)
numqubits(barrier)
numqubits(instruction)
numqubits(circuit)

Number of qubits on which the given operation or instruction is defined.

MimiqCircuitsBase.opname — Function

opname(instruction)
opname(operation)

Returns the name of the given operation in a human readable format.

Gates

MimiqCircuitsBase.Gate — Type

abstract type Gate{N} <: Operation{N}

Supertype for all the N-qubit gates.

Methods

inverse

numqubits

opname

hilbertspacedim

matrix

MimiqCircuitsBase.ParametricGate — Type

abstract type ParametricGate{N}

Supertype for all the parametric N-qubit gates.

Methods

inverse

numqubits

opname

hilbertspacedim

matrix

numparams

parnames

MimiqCircuitsBase.matrix — Function

matrix(gate)

Return the matrix associated to the specified quantum gate.

MimiqCircuitsBase.numparams — Function

numparams(gate)

Number of parameters for the given parametric gate. Zero for non parametric gates.

MimiqCircuitsBase.parnames — Function

parnames(gate)

Name of the parameters allowed for the given gate

Single qubit gates

MimiqCircuitsBase.GateH — Type

struct GateH <: Gate{1}

Single qubit Hadamard gate.

Matrix Representation

\[\operatorname H = \frac{1}{\sqrt{2}} \begin{pmatrix} 1 & 1 \\ 1 & -1 \end{pmatrix}\]

Examples

julia> matrix(GateH())
2×2 Matrix{Float64}:
 0.707107   0.707107
 0.707107  -0.707107

julia> push!(Circuit(), GateH(), 1)
1-qubit circuit with 1 instructions:
└── H @ q1

MimiqCircuitsBase.GateID — Type

struct GateID <: Gate{1}

Single qubit Identity gate

Matrix Representation

\[\operatorname{I} = \begin{pmatrix} 1 & 0 \\ 0 & 1 \end{pmatrix}\]

Examples

julia> matrix(GateID())
2×2 Matrix{Float64}:
 1.0  -0.0
 0.0   1.0

julia> push!(Circuit(), GateID(), 1)
1-qubit circuit with 1 instructions:
└── ID @ q1

MimiqCircuitsBase.GateS — Type

struct GateS <: Gate{1}

Single qubit S gate (or Phase gate).

See also GateSDG

Matrix Representation

\[\operatorname S = \begin{pmatrix} 1 & 0 \\ 0 & i \end{pmatrix}\]

Examples

julia> matrix(GateS())
2×2 Matrix{ComplexF64}:
 1.0+0.0im  0.0+0.0im
 0.0+0.0im  0.0+1.0im

julia> push!(Circuit(), GateS(), 1)
1-qubit circuit with 1 instructions:
└── S @ q1

MimiqCircuitsBase.GateSDG — Type

struct GateSDG <: Gate{1}

Single qubit S-dagger gate (conjugate transpose of the S gate).

See also GateS

Matrix Representation

\[\operatorname S^\dagger = \begin{pmatrix} 1 & 0 \\ 0 & -i \end{pmatrix}\]

Examples

julia> matrix(GateSDG())
2×2 Matrix{ComplexF64}:
 1.0+0.0im  0.0+0.0im
 0.0+0.0im  0.0-1.0im

julia> push!(Circuit(), GateSDG(), 1)
1-qubit circuit with 1 instructions:
└── SDG @ q1

MimiqCircuitsBase.GateSX — Type

struct GateSX <: Gate{1}

Single qubit √X gate.

See also GateSXDG, GateX

Matrix Representation

\[\sqrt{\operatorname{X}} = \frac{1}{2} \begin{pmatrix} 1+i & 1-i \\ 1-i & 1+i \end{pmatrix}\]

Examples

julia> matrix(GateSX())
2×2 Matrix{ComplexF64}:
 0.5+0.5im  0.5-0.5im
 0.5-0.5im  0.5+0.5im

julia> push!(Circuit(), GateSX(), 1)
1-qubit circuit with 1 instructions:
└── SX @ q1

MimiqCircuitsBase.GateSXDG — Type

struct GateSXDG <: Gate{1}

Single qubit √X-dagger gate (conjugate transpose of the √X gate)

See also GateSX, GateX

Matrix Representation

\[\sqrt{\operatorname{X}}^\dagger = \frac{1}{2} \begin{pmatrix} 1-i & 1+i \\ 1+i & 1-i \end{pmatrix}\]

Examples

julia> matrix(GateSXDG())
2×2 Matrix{ComplexF64}:
 0.5-0.5im  0.5+0.5im
 0.5+0.5im  0.5-0.5im

julia> push!(Circuit(), GateSXDG(), 1)
1-qubit circuit with 1 instructions:
└── SXDG @ q1

MimiqCircuitsBase.GateT — Type

struct GateT <: Gate{1}

Single qubit T gate.

See also GateTDG

Matrix Representation

\[\operatorname T = \begin{pmatrix} 1 & 0 \\ 0 & \exp(\frac{i\pi}{4}) \end{pmatrix}\]

Examples

julia> matrix(GateT())
2×2 Matrix{ComplexF64}:
 1.0+0.0im       0.0+0.0im
 0.0+0.0im  0.707107+0.707107im

julia> push!(Circuit(), GateT(), 1)
1-qubit circuit with 1 instructions:
└── T @ q1

MimiqCircuitsBase.GateTDG — Type

struct GateTDG <: Gate{1}

Single qubit T-dagger gate (conjugate transpose of the T gate).

See also GateT

Matrix Representation

\[\operatorname T^\dagger = \begin{pmatrix} 1 & 0 \\ 0 & \exp(\frac{-i\pi}{4}) \end{pmatrix}\]

Examples

julia> matrix(GateTDG())
2×2 Matrix{ComplexF64}:
 1.0+0.0im       0.0+0.0im
 0.0+0.0im  0.707107-0.707107im

julia> push!(Circuit(), GateTDG(), 1)
1-qubit circuit with 1 instructions:
└── TDG @ q1

MimiqCircuitsBase.GateX — Type

struct GateX <: Gate{1}

Single qubit Pauli-X gate.

Matrix Representation

\[\operatorname X = \begin{pmatrix} 0 & 1 \\ 1 & 0 \end{pmatrix}\]

Examples

julia> matrix(GateX())
2×2 Matrix{Float64}:
 0.0   1.0
 1.0  -0.0

julia> push!(Circuit(), GateX(), 1)
1-qubit circuit with 1 instructions:
└── X @ q1

MimiqCircuitsBase.GateY — Type

struct GateY <: Gate{1}

Single qubit Pauli-Y gate.

Matrix Representation

\[\operatorname Y = \begin{pmatrix} 0 & -i \\ i & 0 \end{pmatrix}\]

Examples

julia> matrix(GateY())
2×2 Matrix{ComplexF64}:
 0.0+0.0im  -0.0-1.0im
 0.0+1.0im  -0.0+0.0im

julia> push!(Circuit(), GateY(), 1)
1-qubit circuit with 1 instructions:
└── Y @ q1

MimiqCircuitsBase.GateZ — Type

struct GateZ <: Gate{1}

Single qubit Pauli-Z gate.

Matrix Representation

\[\operatorname Z = \begin{pmatrix} 1 & 0 \\ 0 & -1 \end{pmatrix}\]

Examples

julia> matrix(GateZ())
2×2 Matrix{Float64}:
 1.0   0.0
 0.0  -1.0

julia> push!(Circuit(), GateZ(), 1)
1-qubit circuit with 1 instructions:
└── Z @ q1

MimiqCircuitsBase.GateP — Type

struct GateP <: ParametricGate{1}

Single qubit Phase gate.

Arguments

λ::Float64: Phase angle in radians

Matrix Representation

\[\operatorname P(\lambda) = \begin{pmatrix} 1 & 0 \\ 0 & e^{i\lambda} \end{pmatrix}\]

Examples

julia> matrix(GateP(pi/4))
2×2 Matrix{ComplexF64}:
 1.0+0.0im       0.0+0.0im
 0.0+0.0im  0.707107+0.707107im

julia> push!(Circuit(), GateP(pi/4), 1)
1-qubit circuit with 1 instructions:
└── P(λ=0.7853981633974483) @ q1

MimiqCircuitsBase.GateR — Type

struct GateR <: ParametricGate{1}

Single qubit Rotation gate around the axis cos(ϕ)x + sin(ϕ)y.

Arguments

θ::Float64: Rotation angle in radians

ϕ::Float64: Axis of rotation in radians

Matrix Representation

\[\operatorname R(\theta,\phi) = \begin{pmatrix} \cos\frac{\theta}{2} & -ie^{-i\phi}\sin\frac{\theta}{2} \\ -ie^{-i\phi}\sin\frac{\theta}{2} & \cos\frac{\theta}{2} \end{pmatrix}\]

Examples

julia> matrix(GateR(pi/2,pi/4))
2×2 Matrix{ComplexF64}:
 0.707107+0.0im      -0.5-0.5im
      0.5-0.5im  0.707107+0.0im

julia> push!(Circuit(), GateR(pi/2,pi/4), 1)
1-qubit circuit with 1 instructions:
└── R(θ=1.5707963267948966, ϕ=0.7853981633974483) @ q1

MimiqCircuitsBase.GateRX — Type

struct GateRX <: ParametricGate{1}

Single qubit Rotation-X gate (RX gate)

Arguments

θ::Float64: Rotation angle in radians

Matrix Representation

\[\operatorname{RX}(\theta) = \begin{pmatrix} \cos\frac{\theta}{2} & -i\sin\frac{\theta}{2} \\ -i\sin\frac{\theta}{2} & \cos\frac{\theta}{2} \end{pmatrix}\]

Examples

julia> matrix(GateRX(pi/2))
2×2 Matrix{ComplexF64}:
 0.707107+0.0im           -0.0-0.707107im
      0.0-0.707107im  0.707107+0.0im

julia> push!(Circuit(), GateRX(pi/2), 1)
1-qubit circuit with 1 instructions:
└── RX(θ=1.5707963267948966) @ q1

MimiqCircuitsBase.GateRY — Type

struct GateRY <: ParametricGate{1}

Single qubit Rotation-Y gate (RY gate)

Arguments

θ::Float64: Rotation angle in radians

Matrix Representation

\[\operatorname{RY}(\theta) = \begin{pmatrix} \cos\frac{\theta}{2} & -\sin\frac{\theta}{2} \\ \sin\frac{\theta}{2} & \cos\frac{\theta}{2} \end{pmatrix}\]

Examples

julia> matrix(GateRY(pi/2))
2×2 Matrix{Float64}:
 0.707107  -0.707107
 0.707107   0.707107

julia> push!(Circuit(), GateRY(pi/2), 1)
1-qubit circuit with 1 instructions:
└── RY(θ=1.5707963267948966) @ q1

MimiqCircuitsBase.GateRZ — Type

struct GateRZ <: ParametricGate{1}

Single qubit Rotation-Z gate (RZ gate)

Arguments

λ::Float64: Rotation angle in radians

Matrix Representation

\[\operatorname{RZ}(\lambda) = \begin{pmatrix} e^{-i\frac{\lambda}{2}} & 0 \\ 0 & e^{i\frac{\lambda}{2}} \end{pmatrix}\]

Examples

julia> matrix(GateRZ(pi/2))
2×2 Matrix{ComplexF64}:
 0.707107-0.707107im      -0.0+0.0im
      0.0+0.0im       0.707107+0.707107im

julia> push!(Circuit(), GateRZ(pi/2), 1)
1-qubit circuit with 1 instructions:
└── RZ(λ=1.5707963267948966) @ q1

MimiqCircuitsBase.GateU — Type

struct GateU <: ParametricGate{1}

One qubit generic unitary gate, as defined in OpenQASM 3.0.

Arguments

θ::Float64: Euler angle 1 in radians

ϕ::Float64: Euler angle 2 in radians

λ::Float64: Euler angle 3 in radians

Matrix Representation

\[\operatorname{U}(\theta,\phi,\lambda) = \begin{pmatrix} \cos\frac{\theta}{2} & -e^{i\lambda}\sin\frac{\theta}{2} \\ e^{i\phi}\sin\frac{\theta}{2} & e^{i(\phi+\lambda)}\cos\frac{\theta}{2} \end{pmatrix}\]

Examples

julia> matrix(GateU(pi/3, pi/3, pi/3))
2×2 Matrix{ComplexF64}:
 0.866025+0.0im           -0.25-0.433013im
     0.25+0.433013im  -0.433013+0.75im

julia> push!(Circuit(), GateU(pi/3, pi/3, pi/3), 1)
1-qubit circuit with 1 instructions:
└── U(θ=1.0471975511965976, ϕ=1.0471975511965976, λ=1.0471975511965976) @ q1

MimiqCircuitsBase.GateU1 — Type

struct GateU1 <: ParametricGate{1}

One qubit generic unitary gate u1, as defined in OpenQASM 3.0

Equivalent to GateP

Arguments

λ::Float64: Rotation angle in radians

Matrix Representation

\[\operatorname{U1}(\lambda) = \begin{pmatrix} 1 & 0 \\ 0 & e^{i\lambda} \end{pmatrix}\]

Examples

julia> matrix(GateU1(pi/4))
2×2 Matrix{ComplexF64}:
 1.0+0.0im       0.0+0.0im
 0.0+0.0im  0.707107+0.707107im

julia> push!(Circuit(), GateU1(pi/4), 1)
1-qubit circuit with 1 instructions:
└── U1(λ=0.7853981633974483) @ q1

MimiqCircuitsBase.GateU2 — Type

struct GateU2 <: ParametricGate{1}

One qubit generic unitary gate u2, as defined in OpenQASM 3.0

See also GateU2DG.

Arguments

ϕ:Float64: Rotation angle in radians

λ::Float64: Rotation angle in radians

Examples

julia> matrix(GateU2(pi/2,pi/4))
2×2 Matrix{ComplexF64}:
 0.270598-0.653281im  -0.653281+0.270598im
 0.653281+0.270598im   0.270598+0.653281im

julia> push!(Circuit(), GateU2(pi/4,pi/4), 1)
1-qubit circuit with 1 instructions:
└── U2(ϕ=0.7853981633974483, λ=0.7853981633974483) @ q1

MimiqCircuitsBase.GateU2DG — Type

struct GateU2DG <: ParametricGate{1}

One qubit generic unitary gate u2-dagger, as defined in OpenQASM 3.0 for backwards compatibility

See also GateU2

Arguments

ϕ:Float64: Rotation angle in radians

λ::Float64: Rotation angle in radians

Examples

julia> matrix(GateU2DG(pi/2,pi/4))
2×2 Matrix{ComplexF64}:
  0.270598+0.653281im  0.653281-0.270598im
 -0.653281-0.270598im  0.270598-0.653281im

julia> push!(Circuit(), GateU2DG(pi/2,pi/4), 1)
1-qubit circuit with 1 instructions:
└── U2DG(ϕ=1.5707963267948966, λ=0.7853981633974483) @ q1

MimiqCircuitsBase.GateU3 — Type

struct GateU3 <: ParametricGate{1}

One qubit generic unitary gate u3, as defined in OpenQASM 3.0 for backwards compatibility.

Arguments

θ:Float64: Rotation angle 1 in radians

ϕ:Float64: Rotation angle 2 in radians

λ::Float64: Rotation angle 3 in radians

Examples

julia> matrix(GateU3(pi/2,pi/4,pi/2))
2×2 Matrix{ComplexF64}:
 0.270598-0.653281im  -0.653281-0.270598im
 0.653281-0.270598im   0.270598+0.653281im

julia> push!(Circuit(), GateU3(pi/2, pi/4, pi/2), 1)
1-qubit circuit with 1 instructions:
└── U3(θ=1.5707963267948966, ϕ=0.7853981633974483, λ=1.5707963267948966) @ q1

Two-qubit gates

MimiqCircuitsBase.GateCH — Type

struct GateCH <: Gate{2}

Two qubit Controlled-Hadamard gate.

Matrix Representation

\[\operatorname{CH} = \begin{pmatrix} 1 & 0 & 0 & 0 \\ 0 & 1 & 0 & 0 \\ 0 & 0 & \frac{1}{\sqrt{2}} & \frac{1}{\sqrt{2}} \\ 0 & 0 & \frac{1}{\sqrt{2}} & -\frac{1}{\sqrt{2}} \end{pmatrix}\]

By convention we refer to the first qubit as the control qubit and the second qubit as the target.

Examples

julia> matrix(GateCH())
4×4 Matrix{Float64}:
 1.0  0.0  0.0        0.0
 0.0  1.0  0.0        0.0
 0.0  0.0  0.707107   0.707107
 0.0  0.0  0.707107  -0.707107

julia> push!(Circuit(), GateCH(), 1, 2)
2-qubit circuit with 1 instructions:
└── CH @ q1, q2

MimiqCircuitsBase.GateCS — Type

struct GateCS <: Gate{2}

Two qubit Controlled-S gate.

Matrix Representation

\[\begin{pmatrix} 1 & 0 & 0 & 0 \\ 0 & 1 & 0 & 0 \\ 0 & 0 & 1 & 0 \\ 0 & 0 & 0 & i \end{pmatrix}\]

Examples

julia> matrix(GateCS())
4×4 Matrix{ComplexF64}:
 1.0+0.0im  0.0+0.0im  0.0+0.0im  0.0+0.0im
 0.0+0.0im  1.0+0.0im  0.0+0.0im  0.0+0.0im
 0.0+0.0im  0.0+0.0im  1.0+0.0im  0.0+0.0im
 0.0+0.0im  0.0+0.0im  0.0+0.0im  0.0+1.0im

julia> push!(Circuit(), GateCS(), 1, 2)
2-qubit circuit with 1 instructions:
└── CS @ q1, q2

MimiqCircuitsBase.GateCSDG — Type

struct GateCSDG <: Gate{2}

Two qubit CS-dagger gate.

Matrix Representation

\[\begin{pmatrix} 1 & 0 & 0 & 0 \\ 0 & 1 & 0 & 0 \\ 0 & 0 & 1 & 0 \\ 0 & 0 & 0 & i \end{pmatrix}\]

Examples

julia> matrix(GateCSDG())
4×4 Matrix{ComplexF64}:
 1.0+0.0im  0.0+0.0im  0.0+0.0im  0.0+0.0im
 0.0+0.0im  1.0+0.0im  0.0+0.0im  0.0+0.0im
 0.0+0.0im  0.0+0.0im  1.0+0.0im  0.0+0.0im
 0.0+0.0im  0.0+0.0im  0.0+0.0im  0.0-1.0im

julia> push!(Circuit(), GateCSDG(), 1, 2)
2-qubit circuit with 1 instructions:
└── CSDG @ q1, q2

MimiqCircuitsBase.GateCSX — Type

struct GateCSX <: Gate{2}

Two qubit Controlled-SX gate. (Control on second qubit)

Matrix Representation

\[\begin{pmatrix} 1 & 0 & 0 & 0 \\ 0 & \frac{1+i}{\sqrt{2}} & 0 & \frac{1-i}{\sqrt{2}} \\ 0 & 0 & 1 & 0 \\ 0 & \frac{1-i}{\sqrt{2}} & 0 & \frac{1+i}{\sqrt{2}} \end{pmatrix}\]

Examples

julia> matrix(GateCSX())
4×4 Matrix{ComplexF64}:
 1.0+0.0im  0.0+0.0im  0.0+0.0im  0.0+0.0im
 0.0+0.0im  0.5+0.5im  0.0+0.0im  0.5-0.5im
 0.0+0.0im  0.0+0.0im  1.0+0.0im  0.0+0.0im
 0.0+0.0im  0.5-0.5im  0.0+0.0im  0.5+0.5im

julia> push!(Circuit(), GateCSX(), 1, 2)
2-qubit circuit with 1 instructions:
└── CSX @ q1, q2

MimiqCircuitsBase.GateCSXDG — Type

struct GateCSXDG <: Gate{2}

Two qubit CSX-dagger gate. (Control on second qubit)

Matrix Representation

\[\begin{pmatrix} 1 & 0 & 0 & 0 \\ 0 & \frac{1-i}{\sqrt{2}} & 0 & \frac{1+i}{\sqrt{2}} \\ 0 & 0 & 1 & 0 \\ 0 & \frac{1+i}{\sqrt{2}} & 0 & \frac{1-i}{\sqrt{2}} \end{pmatrix}\]

Examples

julia> matrix(GateCSXDG())
4×4 Matrix{ComplexF64}:
 1.0+0.0im  0.0+0.0im  0.0+0.0im  0.0+0.0im
 0.0+0.0im  0.5-0.5im  0.0+0.0im  0.5+0.5im
 0.0+0.0im  0.0+0.0im  1.0+0.0im  0.0+0.0im
 0.0+0.0im  0.5+0.5im  0.0+0.0im  0.5-0.5im

julia> push!(Circuit(), GateCSXDG(), 1, 2)
2-qubit circuit with 1 instructions:
└── CSXDG @ q1, q2

MimiqCircuitsBase.GateCX — Type

struct GateCX <: Gate{2}

Two qubit Controlled-X gate (or CNOT).

Matrix Representation

\[\operatorname{CX} = \begin{pmatrix} 1 & 0 & 0 & 0 \\ 0 & 1 & 0 & 0 \\ 0 & 0 & 0 & 1 \\ 0 & 0 & 1 & 0 \end{pmatrix}\]

By convention we refer to the first qubit as the control qubit and the second qubit as the target.

Examples

julia> matrix(GateCX())
4×4 Matrix{Float64}:
 1.0  0.0  0.0   0.0
 0.0  1.0  0.0   0.0
 0.0  0.0  0.0   1.0
 0.0  0.0  1.0  -0.0

julia> push!(Circuit(), GateCX(), 1, 2)
2-qubit circuit with 1 instructions:
└── CX @ q1, q2

MimiqCircuitsBase.GateCY — Type

struct GateCY <: Gate{2}

Two qubit Controlled-Y gate.

Matrix Representation

\[\operatorname{CY} = \begin{pmatrix} 1 & 0 & 0 & 0 \\ 0 & 1 & 0 & 0 \\ 0 & 0 & 0 & -i \\ 0 & 0 & i & 0 \end{pmatrix}\]

By convention we refer to the first qubit as the control qubit and the second qubit as the target.

Examples

julia> matrix(GateCY())
4×4 Matrix{ComplexF64}:
 1.0+0.0im  0.0+0.0im  0.0+0.0im   0.0+0.0im
 0.0+0.0im  1.0+0.0im  0.0+0.0im   0.0+0.0im
 0.0+0.0im  0.0+0.0im  0.0+0.0im  -0.0-1.0im
 0.0+0.0im  0.0+0.0im  0.0+1.0im  -0.0+0.0im

julia> push!(Circuit(), GateCY(), 1, 2)
2-qubit circuit with 1 instructions:
└── CY @ q1, q2

MimiqCircuitsBase.GateCZ — Type

struct GateCZ <: Gate{2}

Two qubit Controlled-Z gate.

Matrix Representation

\[\operatorname{CZ} = \begin{pmatrix} 1 & 0 & 0 & 0 \\ 0 & 1 & 0 & 0 \\ 0 & 0 & 1 & 0 \\ 0 & 0 & 0 & -1 \end{pmatrix}\]

By convention we refer to the first qubit as the control qubit and the second qubit as the target.

Examples

julia> matrix(GateCZ())
4×4 Matrix{Float64}:
 1.0  0.0  0.0   0.0
 0.0  1.0  0.0   0.0
 0.0  0.0  1.0   0.0
 0.0  0.0  0.0  -1.0

julia> push!(Circuit(), GateCZ(), 1, 2)
2-qubit circuit with 1 instructions:
└── CZ @ q1, q2

MimiqCircuitsBase.GateDCX — Type

struct GateDCX <: Gate{2}

Two qubit double-CNOT (Control on first qubit and then second) OR DCX gate.

Matrix Representation

\[\begin{pmatrix} 1 & 0 & 0 & 0 \\ 0 & 0 & 0 & 1 \\ 0 & 1 & 0 & 0 \\ 0 & 0 & 1 & 0 \end{pmatrix}\]

Examples

julia> matrix(GateDCX())
4×4 Matrix{Float64}:
 1.0  0.0  0.0  0.0
 0.0  0.0  0.0  1.0
 0.0  1.0  0.0  0.0
 0.0  0.0  1.0  0.0

julia> push!(Circuit(), GateDCX(), 1, 2)
2-qubit circuit with 1 instructions:
└── DCX @ q1, q2

MimiqCircuitsBase.GateDCXDG — Type

struct GateDCXDG <: Gate{2}

Two qubit DCX-dagger gate.

Matrix Representation

\[\begin{pmatrix} 1 & 0 & 0 & 0 \\ 0 & 0 & 1 & 0 \\ 0 & 0 & 0 & 1 \\ 0 & 1 & 0 & 0 \end{pmatrix}\]

Examples

julia> matrix(GateDCXDG())
4×4 Matrix{Float64}:
 1.0  0.0  0.0  0.0
 0.0  0.0  1.0  0.0
 0.0  0.0  0.0  1.0
 0.0  1.0  0.0  0.0

julia> push!(Circuit(), GateDCXDG(), 1, 2)
2-qubit circuit with 1 instructions:
└── DCXDG @ q1, q2

MimiqCircuitsBase.GateECR — Type

struct GateECR <: Gate{2}

Two qubit ECR echo gate.

Matrix Representation

\[\begin{pmatrix} 0 & \frac{1}{\sqrt{2}} \ & 0 & \frac{i}{\sqrt{2}} \\ \frac{1}{\sqrt{2}} & 0 & \frac{-i}{\\sqrt{2}} & 0 \\ 0 & \frac{i}{\\sqrt{2}} & 0 & \frac{i}{\sqrt{2}} \\ \frac{-i}{\sqrt{2}} & 0 & \frac{1}{\sqrt{2}} & 0 \end{pmatrix}\]

Examples

julia> matrix(GateCSX())
4×4 Matrix{ComplexF64}:
 1.0+0.0im  0.0+0.0im  0.0+0.0im  0.0+0.0im
 0.0+0.0im  0.5+0.5im  0.0+0.0im  0.5-0.5im
 0.0+0.0im  0.0+0.0im  1.0+0.0im  0.0+0.0im
 0.0+0.0im  0.5-0.5im  0.0+0.0im  0.5+0.5im

julia> push!(Circuit(), GateCSX(), 1, 2)
2-qubit circuit with 1 instructions:
└── CSX @ q1, q2

MimiqCircuitsBase.GateISWAP — Type

struct GateISWAP <: Gate{2}

Two qubit ISWAP gate.

See also GateISWAPDG, GateSWAP.

Matrix Representation

\[\operatorname{ISWAP} = \frac{1}{\sqrt{2}} \begin{pmatrix} 1 & 0 & 0 & 0 \\ 0 & 0 & i & 0 \\ 0 & i & 0 & 0 \\ 0 & 0 & 0 & 1 \end{pmatrix}\]

Examples

julia> matrix(GateISWAP())
4×4 Matrix{ComplexF64}:
 1.0+0.0im  0.0+0.0im  0.0+0.0im  0.0+0.0im
 0.0+0.0im  0.0+0.0im  0.0+1.0im  0.0+0.0im
 0.0+0.0im  0.0+1.0im  0.0+0.0im  0.0+0.0im
 0.0+0.0im  0.0+0.0im  0.0+0.0im  1.0+0.0im

julia> push!(Circuit(), GateISWAP(), 1, 2)
2-qubit circuit with 1 instructions:
└── ISWAP @ q1, q2

MimiqCircuitsBase.GateISWAPDG — Type

struct GateISWAPDG <: Gate{2}

Two qubit ISWAP-dagger gate (conjugate transpose of ISWAP)

See also GateISWAP, GateSWAP

Matrix Representation

\[\operatorname{ISWAP}^\dagger = \frac{1}{\sqrt{2}} \begin{pmatrix} 1 & 0 & 0 & 0 \\ 0 & 0 & -i & 0 \\ 0 & -i & 0 & 0 \\ 0 & 0 & 0 & 1 \end{pmatrix}\]

Examples

julia> matrix(GateISWAPDG())
4×4 Matrix{ComplexF64}:
 1.0+0.0im  0.0+0.0im  0.0+0.0im  0.0+0.0im
 0.0+0.0im  0.0+0.0im  0.0-1.0im  0.0+0.0im
 0.0+0.0im  0.0-1.0im  0.0+0.0im  0.0+0.0im
 0.0+0.0im  0.0+0.0im  0.0+0.0im  1.0+0.0im

julia> push!(Circuit(), GateISWAPDG(), 1, 2)
2-qubit circuit with 1 instructions:
└── ISWAPDG @ q1, q2

MimiqCircuitsBase.GateSWAP — Type

struct GateSWAP <: Gate{2}

Two qubit SWAP gate.

See also GateISWAP

Matrix Representation

\[\operatorname{SWAP} = \frac{1}{\sqrt{2}} \begin{pmatrix} 1 & 0 & 0 & 0 \\ 0 & 0 & 1 & 0 \\ 0 & 1 & 0 & 0 \\ 0 & 0 & 0 & 1 \end{pmatrix}\]

Examples

julia> matrix(GateSWAP())
4×4 Matrix{Float64}:
 1.0  0.0  0.0  0.0
 0.0  0.0  1.0  0.0
 0.0  1.0  0.0  0.0
 0.0  0.0  0.0  1.0

julia> push!(Circuit(), GateSWAP(), 1, 2)
2-qubit circuit with 1 instructions:
└── SWAP @ q1, q2

MimiqCircuitsBase.GateCP — Type

struct GateCP <: ParametricGate{2}

Two qubit Controlled-Phase gate

Arguments

λ::Float64: Phase angle in radians

Matrix Representation

\[\operatorname{CP}(\lambda) = \begin{pmatrix} 1 & 0 & 0 & 0 \\ 0 & 1 & 0 & 0 \\ 0 & 0 & 1 & 0 \\ 0 & 0 & 0 & e^{i\lambda} \end{pmatrix}\]

By convention we refer to the first qubit as the control qubit and the second qubit as the target.

Examples

julia> matrix(GateCP(pi/4))
4×4 Matrix{ComplexF64}:
 1.0+0.0im  0.0+0.0im  0.0+0.0im       0.0+0.0im
 0.0+0.0im  1.0+0.0im  0.0+0.0im       0.0+0.0im
 0.0+0.0im  0.0+0.0im  1.0+0.0im       0.0+0.0im
 0.0+0.0im  0.0+0.0im  0.0+0.0im  0.707107+0.707107im

julia> push!(Circuit(), GateCP(pi/4), 1, 2)
2-qubit circuit with 1 instructions:
└── CP(λ=0.7853981633974483) @ q1, q2

MimiqCircuitsBase.GateCR — Type

struct GateCR <: ParametricGate{2}

Two qubit Controlled-R gate.

Arguments

θ::Float64: Rotation angle in radians

ϕ::Float64: The phase angle in radians.

Matrix Representation

\[\operatorname{CR}(\theta, \phi) = \begin{pmatrix} 1 & 0 & 0 & 0 \\ 0 & 1 & 0 & 0 \\ 0 & 0 & \cos\frac{\theta}{2} & -ie^{-i\phi}\sin\\frac{\theta}{2} \\ 0 & 0 & -ie^{i\phi}\sin\frac{\theta}{2} & \cos\frac{\theta}{2} \end{pmatrix}\]

Examples

julia> matrix(GateCR(pi,-pi))
4×4 Matrix{ComplexF64}:
 1.0+0.0im  0.0+0.0im           0.0+0.0im          0.0+0.0im
 0.0+0.0im  1.0+0.0im           0.0+0.0im          0.0+0.0im
 0.0+0.0im  0.0+0.0im   6.12323e-17+0.0im  1.22465e-16+1.0im
 0.0+0.0im  0.0+0.0im  -1.22465e-16+1.0im  6.12323e-17+0.0im

julia> push!(Circuit(), GateCR(pi,-pi), 1, 2)
2-qubit circuit with 1 instructions:
└── CR(θ=3.141592653589793, ϕ=-3.141592653589793) @ q1, q2

MimiqCircuitsBase.GateCRX — Type

struct GateCRX <: ParametricGate{2}

Two qubit Controlled-RX gate

Arguments

θ::Float64: Rotation angle in radians

Matrix Representation

\[\operatorname{CRX}(\theta) = \begin{pmatrix} 1 & 0 & 0 & 0 \\ 0 & 1 & 0 & 0 \\ 0 & 0 & \cos\frac{\theta}{2} & -i\sin\frac{\theta}{2} \\ 0 & 0 & -i\sin\frac{\theta}{2} & \cos\frac{\theta}{2} \end{pmatrix}\]

By convention we refer to the first qubit as the control qubit and the second qubit as the target.

Examples

julia> matrix(GateCRX(pi/2))
4×4 Matrix{ComplexF64}:
 1.0+0.0im  0.0+0.0im       0.0+0.0im            0.0+0.0im
 0.0+0.0im  1.0+0.0im       0.0+0.0im            0.0+0.0im
 0.0+0.0im  0.0+0.0im  0.707107+0.0im           -0.0-0.707107im
 0.0+0.0im  0.0+0.0im       0.0-0.707107im  0.707107+0.0im

julia> push!(Circuit(), GateCRX(pi/2), 1, 2)
2-qubit circuit with 1 instructions:
└── CRX(θ=1.5707963267948966) @ q1, q2

MimiqCircuitsBase.GateCRY — Type

struct GateCRY <: ParametricGate{2}

Two qubit Controlled-RY gate

Arguments

θ::Float64: Rotation angle in radians

Matrix Representation

\[\operatorname{CRY}(\theta) = \begin{pmatrix} 1 & 0 & 0 & 0 \\ 0 & 1 & 0 & 0 \\ 0 & 0 & \cos\frac{\theta}{2} & -\sin\frac{\theta}{2} \\ 0 & 0 & \sin\frac{\theta}{2} & \cos\frac{\theta}{2} \end{pmatrix}\]

By convention we refer to the first qubit as the control qubit and the second qubit as the target.

Examples

julia> matrix(GateCRY(pi/2))
4×4 Matrix{Float64}:
 1.0  0.0  0.0        0.0
 0.0  1.0  0.0        0.0
 0.0  0.0  0.707107  -0.707107
 0.0  0.0  0.707107   0.707107

julia> push!(Circuit(), GateCRY(pi/2), 1, 2)
2-qubit circuit with 1 instructions:
└── CRY(θ=1.5707963267948966) @ q1, q2

MimiqCircuitsBase.GateCRZ — Type

struct GateCRZ <: ParametricGate{2}

Two qubit Controlled-RZ gate

Arguments

λ::Float64: Rotation angle in radians

Matrix Representation

\[\operatorname{CRZ}(\lambda) = \begin{pmatrix} 1 & 0 & 0 & 0 \\ 0 & 1 & 0 & 0 \\ 0 & 0 & e^{-i\frac{\lambda}{2}} & 0 \\ 0 & 0 & 0 & e^{i\frac{\lambda}{2}} \end{pmatrix}\]

By convention we refer to the first qubit as the control qubit and the second qubit as the target.

Examples

julia> matrix(GateCRZ(pi/2))
4×4 Matrix{ComplexF64}:
 1.0+0.0im  0.0+0.0im       0.0+0.0im            0.0+0.0im
 0.0+0.0im  1.0+0.0im       0.0+0.0im            0.0+0.0im
 0.0+0.0im  0.0+0.0im  0.707107-0.707107im      -0.0+0.0im
 0.0+0.0im  0.0+0.0im       0.0+0.0im       0.707107+0.707107im

julia> push!(Circuit(), GateCRZ(pi/2), 1, 2)
2-qubit circuit with 1 instructions:
└── CRZ(λ=1.5707963267948966) @ q1, q2

MimiqCircuitsBase.GateCU — Type

struct GateCU <: ParametricGate{2}

Two qubit generic unitary gate, equivalent to the qiskit CUGate https://qiskit.org/documentation/stubs/qiskit.circuit.library.CUGate.html

Arguments

θ::Float64: Euler angle 1 in radians

ϕ::Float64: Euler angle 2 in radians

λ::Float64: Euler angle 3 in radians

γ::Float64: Global phase of the U gate

Matrix Representation

\[\operatorname{CU}(\theta,\phi,\lambda,\gamma) = \begin{pmatrix} 1 & 0 & 0 & 0 \\ 0 & 1 & 0 & 0 \\ 0 & 0 & e^{i\gamma}\cos\frac{\theta}{2} & -e^{i(\gamma+\lambda)}\sin\frac{\theta}{2} \\ 0 & 0 & e^{i(\gamma+\phi)}\sin\frac{\theta}{2} & e^{i(\gamma+\phi+\lambda)}\cos\frac{\theta}{2} \end{pmatrix}\]

By convention we refer to the first qubit as the control qubit and the second qubit as the target.

Examples

julia> matrix(GateCU(pi/3, pi/3, pi/3, 0))
4×4 Matrix{ComplexF64}:
 1.0+0.0im  0.0+0.0im       0.0+0.0im             0.0+0.0im
 0.0+0.0im  1.0+0.0im       0.0+0.0im             0.0+0.0im
 0.0+0.0im  0.0+0.0im  0.866025+0.0im           -0.25-0.433013im
 0.0+0.0im  0.0+0.0im      0.25+0.433013im  -0.433013+0.75im

julia> push!(Circuit(), GateCU(pi/3, pi/3, pi/3, 0), 1, 2)
2-qubit circuit with 1 instructions:
└── CU(θ=1.0471975511965976, ϕ=1.0471975511965976, λ=1.0471975511965976, γ=0.0) @ q1, q2

MimiqCircuitsBase.GateRXX — Type

struct GateRXX <: ParametricGate{2}

Two qubit RXX gate.

Arguments

θ::Float64: The angle in radians

Matrix Representation

\[\operatorname{RXX}(\theta) = \begin{pmatrix} \cos(\\frac{\theta}{2}) & 0 & 0 & -i\sin(\\frac{\theta}{2}) \\ 0 & \cos(\frac{\theta}{2}) & -i\sin(\frac{\theta}{2}) & 0 \\ 0 & -i\sin(\frac{\theta}{2}) & \cos(\frac{\theta}{2}) & 0 \\ -i\sin(\frac{\theta}{2}) & 0 & 0 & \cos(\frac{\theta}{2}) \end{pmatrix}\]

Examples

julia> matrix(GateRXX(π/4))
4×4 Matrix{ComplexF64}:
 0.92388+0.0im           0.0+0.0im       …      0.0-0.382683im
     0.0+0.0im       0.92388+0.0im              0.0+0.0im
     0.0+0.0im           0.0-0.382683im         0.0+0.0im
     0.0-0.382683im      0.0+0.0im          0.92388+0.0im

julia> push!(Circuit(), GateRXX(π), 1, 2)
2-qubit circuit with 1 instructions:
└── RXX(θ=3.141592653589793) @ q1, q2

MimiqCircuitsBase.GateRYY — Type

struct GateRYY <: ParametricGate{2}

Two qubit RYY gate.

Arguments

θ::Float64: The angle in radians

Matrix Representation

\[\operatorname{RYY}(\theta) = \begin{pmatrix} \cos(\frac{\theta}{2}) & 0 & 0 & i\sin(\frac{\theta}{2}) \\ 0 & \cos(\frac{\theta}{2}) & -i\sin(\frac{\theta}{2}) & 0 \\ 0 & -i\sin(\\frac{\theta}{2}) & \cos(\frac{\theta}{2}) & 0 \\ i\sin(\frac{\theta}{2}) & 0 & 0 & \cos(\frac{\theta}{2}) \end{pmatrix}\]

Examples

julia> matrix(GateRYY(π/4))
4×4 Matrix{ComplexF64}:
 0.92388+0.0im           0.0+0.0im       …      0.0+0.382683im
     0.0+0.0im       0.92388+0.0im              0.0+0.0im
     0.0+0.0im           0.0-0.382683im         0.0+0.0im
     0.0+0.382683im      0.0+0.0im          0.92388+0.0im

julia> push!(Circuit(), GateRYY(π), 1, 2)
2-qubit circuit with 1 instructions:
└── RYY(θ=3.141592653589793) @ q1, q2

MimiqCircuitsBase.GateRZZ — Type

struct GateRZZ <: ParametricGate{2}

Two qubit RZZ gate.

Arguments

θ::Float64: The angle in radians

Matrix Representation

\[\operatorname{RZZ}(\theta) = \begin{pmatrix} e^{-i\frac{\theta}{2}} & 0 & 0 & 0 \\ 0 & e^{i\frac{\theta}{2}} & 0 & 0 \\ 0 & 0 & e^{i\frac{\theta}{2}} & 0 \\ 0 & 0 & 0 & e^{-i\frac{\theta}{2}} \end{pmatrix}\]

Examples

julia> matrix(GateRZZ(π/4))
4×4 Matrix{ComplexF64}:
 0.92388-0.382683im      0.0+0.0im       …      0.0+0.0im
     0.0+0.0im       0.92388+0.382683im         0.0+0.0im
     0.0+0.0im           0.0+0.0im              0.0+0.0im
     0.0+0.0im           0.0+0.0im          0.92388-0.382683im

julia> push!(Circuit(), GateRZZ(π), 1, 2)
2-qubit circuit with 1 instructions:
└── RZZ(θ=3.141592653589793) @ q1, q2

MimiqCircuitsBase.GateXXminusYY — Type

struct GateXXminusYY <: ParametricGate{2}

Two qubit XXminusYY gate.

Arguments

θ::Float64: The angle in radians.

β::Float64: The phase angle in radians.

Matrix Representation

\[\operatorname{XXminusYY}(\theta, \beta) = \begin{pmatrix} \cos(\frac{\theta}{2}) & 0 & 0 & -i\sin(\frac{\theta}{2})e^{-i\beta} \\ 0 & 1 & 0 & 0 \\ 0 & 0 & 1 & 0 \\ -i\sin(\frac{\theta}{2})e^{i\beta} & 0 & 0 & \cos(\frac{\theta}{2}) \end{pmatrix}\]

Examples

julia> matrix(GateXXminusYY(π/2,π/2))
4×4 Matrix{ComplexF64}:
 0.707107+0.0im          0.0+0.0im  0.0+0.0im  -0.707107-4.32978e-17im
      0.0+0.0im          1.0+0.0im  0.0+0.0im        0.0+0.0im
      0.0+0.0im          0.0+0.0im  1.0+0.0im        0.0+0.0im
 0.707107-4.32978e-17im  0.0+0.0im  0.0+0.0im   0.707107+0.0im

julia> push!(Circuit(), GateXXminusYY(π,π), 1, 2)
2-qubit circuit with 1 instructions:
└── XXminusYY(θ=3.141592653589793, β=3.141592653589793) @ q1, q2

MimiqCircuitsBase.GateXXplusYY — Type

struct GateXXplusYY <: ParametricGate{2}

Two qubit XXplusYY gate.

Arguments

θ::Float64: The angle in radians.

β::Float64: The phase angle in radians.

Matrix Representation

\[\operatorname{XXplusYY}(\theta, \beta) = \begin{pmatrix} 1 & 0 & 0 & 0 \\ 0 & \cos(\frac{\theta}{2}) & -i\sin(\frac{\theta}{2})e^{-i\beta} & 0 \\ 0 & -i\sin(\\frac{theta}{2})e^{i\beta} & \cos(\frac{\theta}{2}) & 0 \\ 0 & 0 & 0 & 1 \end{pmatrix}\]

Examples

julia> matrix(GateXXplusYY(π/2,π/2))
4×4 Matrix{ComplexF64}:
 1.0+0.0im       0.0+0.0im                0.0+0.0im          0.0+0.0im
 0.0+0.0im  0.707107+0.0im          -0.707107-4.32978e-17im  0.0+0.0im
 0.0+0.0im  0.707107-4.32978e-17im   0.707107+0.0im          0.0+0.0im
 0.0+0.0im       0.0+0.0im                0.0+0.0im          1.0+0.0im

julia> push!(Circuit(), GateXXplusYY(π,π), 1, 2)
2-qubit circuit with 1 instructions:
└── XXplusYY(θ=3.141592653589793, β=3.141592653589793) @ q1, q2

Multi-qubit

MimiqCircuitsBase.GateCCX — Type

struct GateCCX <: Gate{3}

C₂X (or C₂NOT) 3-qubits gate. Where the first two qubits are used as controls.

MimiqCircuitsBase.GateCSWAP — Type

struct GateCSWAP <: Gate{3} end

3-qubits control SWAP gate where the first qubit is the control.

Custom gates

MimiqCircuitsBase.GateCustom — Type

struct GateCustom{N,T} <: Gate{N}

N qubit gate specified by a $2^N \times 2^N$ matrix with elements of type T.

Use this to construct your own gates based on unitary matrices. Currently only N=1,2 (M=2,4) are recognised.

MIMIQ uses textbook convention for specifying gates.

One qubit gate matrices are defined in the basis $|0\rangle$, $|1\rangle$ e.g.,

\[\operatorname{Z} = \begin{pmatrix} 1&0\\ 0&-1 \end{pmatrix}\]

Two qubit gate matrices are defined in the basis $|00\rangle$, $|01\rangle$>, $|10\rangle$, $|11\rangle$ where the left-most qubit is the first to appear in the target list e.g.,

\[\operatorname{CNOT} = \begin{pmatrix} 1&0&0&0\\ 0&1&0&0\\ 0&0&0&1\\ 0&0&1&0 \end{pmatrix}\]

julia> CNOT = [1 0 0 0; 0 1 0 0; 0 0 0 1; 0 0 1 0]
4×4 Matrix{Int64}:
 1  0  0  0
 0  1  0  0
 0  0  0  1
 0  0  1  0

julia> # CNOT gate with control on q1 and target on q2

julia> Instruction(GateCustom(CNOT), 1, 2)
GateCustom([1 0 0 0; 0 1 0 0; 0 0 0 1; 0 0 1 0]) @ q1, q2

# Examples

jldoctest julia> g = GateCustom([1 0; 0 1]) Custom([1.0 0.0; 0.0 1.0])

julia> push!(Circuit(), g, 1) 1-qubit circuit with 1 instructions: └── Custom([1.0 0.0; 0.0 1.0]) @ q1 ```

Non-unitary operations

MimiqCircuitsBase.Reset — Type

struct Reset <: Operation{1}

Quantum operation that resets the status of one qubit to the $\ket{0}$ state.

Examples

julia> push!(Circuit(), Reset(), 1)
1-qubit circuit with 1 instructions:
└── Reset @ q1

MimiqCircuitsBase.Measure — Type

struct Measure <: Operation{1}

Single qubit measurement operation in the computational basis

This operation is non-reversible

Examples

Measure project the qubit state and optionally store the result of the measurement for that qubit in a classical register.

To just apply the measurement on qubit 1 and discard the result, do:

julia> push!(Circuit(), Measure(), 1,1)
1-qubit circuit with 1 instructions:
└── Measure @ q1, c1

No-ops

MimiqCircuitsBase.Barrier — Type

struct Barrier <: Operation{1}

A barrier is a special operation that does not affect the quantum state or the execution of a circuit, but it prevents compression or optimization operation from being applied across it.

Examples

julia> push!(Circuit(), Barrier(), 1)
1-qubit circuit with 1 instructions:
└── Barrier @ q1

julia> push!(Circuit(), Barrier(4), 1:4...)
4-qubit circuit with 1 instructions:
└── Barrier @ q1, q2, q3, q4

julia> push!(Circuit(), Barrier(), 1:4)
4-qubit circuit with 4 instructions:
├── Barrier @ q1
├── Barrier @ q2
├── Barrier @ q3
└── Barrier @ q4

Composite operations

MimiqCircuitsBase.IfStatement — Type

struct IfStatement{N,M} <: Operation{N,M}

MimiqCircuitsBase.Control — Type

struct Control{N,M,L,T<:Operation{M,0}} <: Operation{L,0} end

MimiqCircuitsBase.Parallel — Type

struct Parallel{N,M,L,T<:Operation{M,0}} <: Operation{L,0} end

Bit States

MimiqCircuitsBase.BitState — Type

struct BitState

Representation of the quantum state of a quantum register with definite values for each qubit.

Examples

julia> BitState(16)
16-qubit BitState with 0 non-zero qubits:
└── |0000000000000000⟩

julia> bs = BitState(16, [1,2,3,4])
16-qubit BitState with 4 non-zero qubits:
├── |1111000000000000⟩
└── non-zero qubits: [1, 2, 3, 4]

julia> bs[10] = 1
1

julia> bs
16-qubit BitState with 5 non-zero qubits:
├── |1111000001000000⟩
└── non-zero qubits: [1, 2, 3, 4, 10]

julia> c = Circuit()
empty circuit

julia> push!(c, GateX(), 8)
8-qubit circuit with 1 instructions:
└── X @ q8

julia> BitState(c, [1,3,5,8])
8-qubit BitState with 4 non-zero qubits:
├── |10101001⟩
└── non-zero qubits: [1, 3, 5, 8]

julia> bitstate_to_integer(bs)
527

julia> typeof(ans)
BigInt

julia> bitstate_to_integer(bs, Int64)
527

julia> typeof(ans)
Int64

There are many different ways to get bit states:

julia> bs = BitState(30, 2344574)
30-qubit BitState with 13 non-zero qubits:
├── |011111100110001111000100000000⟩
└── non-zero qubits: [2, 3, 4, 5, 6, 7, 10, 11, 15, 16, 17, 18, 22]

julia> ones(BitState, 10) # or also trues(BitState, 10)
10-qubit BitState with 10 non-zero qubits:
├── |1111111111⟩
└── non-zero qubits: [1, 2, 3, 4, 5, 6, 7, 8, 9, 10]

julia> zeros(BitState, 10) # or also falses(BitState, 10)
10-qubit BitState with 0 non-zero qubits:
└── |0000000000⟩

julia> BitState(16) do i
           iseven(i)
       end
16-qubit BitState with 8 non-zero qubits:
├── |0101010101010101⟩
└── non-zero qubits: [2, 4, 6, 8, 10, 12, 14, 16]

MimiqCircuitsBase.bits — Method

bits(bitstate)

Get the underlying bit array of a bit state.

MimiqCircuitsBase.bitstate_to_index — Method

bitstate_to_index(bitstate)

Convert a bit state into the corresponding index.

This is useful for indexing, for example, a vector of states.

MimiqCircuitsBase.bitstate_to_integer — Method

bitstate_to_integer(bitstate[, T])

Convert a bit state into its corresponding integer.

MimiqCircuitsBase.nonzeros — Method

nonzeros(bitstate)

Return the indices of the non-zero qubits in a bit state.

MimiqCircuitsBase.to01 — Method

to01(bitstate[, endianess=:big])

Converts a BitState into a string of 0 and 1 characters. Optionally endianess can be specified, which can be either :big or :little.

Examples

julia> to01(bs"10011")
"10011"

julia> to01(bs"10011"; endianess=:big)
"10011"

julia> to01(bs"10011"; endianess=:little)
"11001"

MimiqCircuitsBase.@bs_str — Macro

macro bs_str(s)

Convert a string into a bit state.

Examples

julia> bs"101011"
6-qubit BitState with 4 non-zero qubits:
├── |101011⟩
└── non-zero qubits: [1, 3, 5, 6]